passengerCOVIDscan/requirements.txt
absl-py==0.9.0
astor==0.8.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dj-static==0.0.6
Django==2.2
gast==0.2.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
gunicorn==19.1.1
h5py==2.10.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.0
numpy==1.18.1
oauthlib==3.1.0
opencv-python
opt-einsum==3.2.1
Pillow==6.2.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
pyzbar==0.1.8
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
tensorboard==2.1.1
tensorflow-estimator==2.1.0
termcolor==1.1.0
torchvision==0.2.2
tqdm==4.19.9
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
